FOLLOW-UP: Re-added phd image, piechart size smaller
Added old phd image as current script functions needed it else the script crashes with errors. 

Also resized the picture slightly. The picture is generated at a size and isn't  made again by DOM when resizing screen. It is thus not responsive but it should be alright with this size.